% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_confusion_data.R
\name{simulate_confusion_data}
\alias{simulate_confusion_data}
\title{Simulate correct and incorrect observations of diagnostic test result}
\usage{
simulate_confusion_data(TP, FP, FN, TN, R = 10^5)
}
\arguments{
\item{TP}{True positives. Must be a numeric vector of length 1.}

\item{FP}{False positives. Must be a numeric vector of length 1.}

\item{FN}{False negatives. Must be a numeric vector of length 1.}

\item{TN}{True negatives. Must be a numeric vector of length 1.}

\item{R}{Number of Monte Carlo simulations. Numeric vector of length 1.}
}
\value{
An n x 4 matrix where n is the number of simulations.
}
\description{
This function uses Monte Carlo simulated
diagnostic observations based on the assumption that the
number of true positives and true negatives are independent
binomial random variables, using the sample sensitivity and specificity
as estimates of the true proportions respectively.
}
\examples{
# Using the results from Anu Kantele et al. (2021) as an example
library(epimetrics)
TP <- 392
FP <- 75
FN <- 35
TN <- 807
sim_cm <- simulate_confusion_data(TP, FP, FN, TN, R = 1000)
sim_cm
ppv(x = sim_cm)
TP/(TP + FP)
hist(ppv(x = sim_cm)[,1])
}
