% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wembc_ci.R
\name{wembc_ci}
\alias{wembc_ci}
\title{Wilson-score interval with exact method
boundary correction for extreme cases}
\usage{
wembc_ci(x, n, alpha = 0.05)
}
\arguments{
\item{x}{Numeric vector containing the number of successes.}

\item{n}{Numeric vector containing the number of trials.}

\item{alpha}{Numeric vector containing the level of significance.}
}
\value{
An n x 3 \code{matrix} with the columns:
\item{estimate}{point-estimate of the data.}
\item{conf.low}{lower confidence bound of the estimate.}
\item{conf.high}{upper confidence bound of the estimate.}
}
\description{
This is an efficient function to compute a
binomial proportion with confidence intervals using the
wilson-score interval method, with a boundary correction for when the
number of successes are either 0, or equal the number of trials.
}
\examples{
library(epimetrics)
wembc_ci(0, 0:10)
wembc_ci(0:10, 0:10)
combs <- expand.grid(0:999, 1000:2000)
wembc_ci(combs[,1], combs[,2])
}
\references{
Sakthivel Sivam, S. M.
"Everything or Nothing-A Better Confidence Intervals for Binomial
Proportion in Clinical Trial Data Analysis."
Sakthivel Sivam, Quartesian LLC,
Princeton, New Jersey Subbiah Meenakshisundaram, L. N Government College,
Ponneri, India 2016 (2014).
\url{https://www.lexjansen.com/pharmasug/2016/SP/PharmaSUG-2016-SP08.pdf}
}
