% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix}
\usage{
confusion_matrix(outcome, prediction)
}
\arguments{
\item{outcome}{Vector of outcome/response/actual classes. e.g. 0, 1.}

\item{prediction}{Vector of predicted classes. e.g. "A", "B".}
}
\description{
Wrapper around \code{table()} that
creates an n x m confusion matrix of the number of
correct and incorrect classes of a multi-class variable where
n is the number of distinct predicted classes and m is the
number of distinct outcome classes.
The only difference between \code{confusion_matrix()} and \code{table()}
is that it sorts the rows and columns
of the predicted and actual classes respectively in descending order,
similar to the way that confusion matrices are usually presented.
}
\examples{
\dontrun{
library(epimetrics)
library(MASS)
library(randomForest)

pima_train <- Pima.tr # Training data set
pima_test <- Pima.te # Test data set

rf_train <- randomForest(type ~ ., data = pima_train)
rf_predictions <- predict(rf_train, newdata = pima_test)
outcomes <- pima_test$type
cm <- confusion_matrix(outcomes, rf_predictions)
cm
diagnostic_metrics(cm)
}
}
